+++ packages/provider/src/tests/unit/util.unit.test.ts
@@ -22,6 +22,7 @@ import { Address4, Address6 } from "ip-address";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import {
 	checkIfTaskIsRunning,
+	deepValidateIpAddress,
 	getIPAddress,
 	validateIpAddress,
 } from "../../util.js";