+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prosopo/client-example-server
 
+## 2.7.30
+### Patch Changes
+
+- 0d1a33e: Adding ipcomparison service with user features
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types@3.1.4
+  - @prosopo/keyring@2.8.18
+  - @prosopo/common@3.1.11
+  - @prosopo/dotenv@3.0.16
+  - @prosopo/server@2.9.26
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
 ## 2.7.29
 ### Patch Changes