+++ demos/provider-mock/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prosopo/provider-mock
 
+## 2.8.8
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types-database@3.1.5
+  - @prosopo/provider@3.8.4
+  - @prosopo/locale@3.1.11
+  - @prosopo/types@3.1.4
+  - @prosopo/api-express-router@3.0.16
+  - @prosopo/database@3.2.4
+  - @prosopo/common@3.1.11
+  - @prosopo/dotenv@3.0.16
+  - @prosopo/config@3.1.11
+
 ## 2.8.7
 ### Patch Changes